+++ Core/Runtime/UnityExtensions/HisaCat/UnityExtensions/Components/DestroySelfOnAwake.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 namespace HisaCat
 {
-    [DefaultExecutionOrder(short.MinValue)]
+    [DefaultExecutionOrder(int.MinValue)]
     public class DestroySelfOnAwake : MonoBehaviour
     {
         private void Awake() => Destroy(this.gameObject);
